requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
langchain>=0.0.350
openai>=1.0.0
httpx>=0.25.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0
python-dotenv>=1.0.0
loguru>=0.7.0
pyyaml>=6.0.0
python-telegram-bot>=20.0.0
openai-whisper>=20231117
TTS>=0.22.0
piper-tts>=1.2.0
